Renamed ui folder to components
