+++ src/components/AnimatedDescriptions.jsx
