+++ src/components/Profile.jsx
@@ -1,4 +1,4 @@
-import { Button, Dropdown } from "antd";
+import { Dropdown } from "antd";
 import UserHeader from "./UserHeader";
 import { useNavigate } from "react-router-dom";
